+++ src/oblivion_orb.rs
@@ -8,7 +8,7 @@ pub struct OblivionOrb {
     price: usize,
     magic_power: i32,
     effect_heal_reduce: usize,
-    effect_duration_seconds: usize,
+    effect_duration_seconds: f64,
 }
 
 impl Default for OblivionOrb {
@@ -17,7 +17,7 @@ impl Default for OblivionOrb {
             price: 500,
             magic_power: 50,
             effect_heal_reduce: 25,
-            effect_duration_seconds: 2,
+            effect_duration_seconds: 2.0,
         }
     }
 }
@@ -95,7 +95,7 @@ impl ModItemInfo for OblivionOrb {
                 target,
                 BuffState {
                     duration: BuffType::Time {
-                        tick: self.effect_duration_seconds * 60,
+                        tick: (self.effect_duration_seconds * 60.0) as usize,
                     },
                     heal_reduce: self.effect_heal_reduce,
                     name: ArrayString::try_from("25_percent_heal_cut").unwrap(),
